backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
django-prometheus
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt
django-allauth
bcrypt
django-cors-headers==4.3
groq>=0.9
langdetect==1.0.9
requests>=2.32.0
coverage==7.6.1